requirements.txt
numpy
pandas
matplotlib
seaborn
torch
transformers
tensorboard
torchtext
openpyxl
gensim
joblib
python-box
ensure
torchinfo
scikit-learn
optuna
xgboost
lightgbm